Run free automated tests on Linux and MacOSX using Travis CI
Travis CI can be used to do automated testing on all pull requests before they are reviewed and merged. This service is free to open source projects like this one.  For this to work, @mrJean1 would need to turn on the switch for this repo at https://www.travis-ci.org/mrJean1 (login via GitHub credentials)