Hartzell and height4
New function and Cartesian and LatLon method hartzell to compute the intersection of an earth model with a Line-Of-Sight from a Point-Of-View in space.

New Ellipsoid, Cartesian and LatLon method height4 to compute the projection and height of a point above or below the surface of an earth model.

Upgrade to macOS 11.5.1 for building and testing.