More __special__ methods
Upgrade to macOS 12.1 Monterey and Python 3.10.1 for testing and building.

Additional __special__ methods for vectorial classes as Cartesian, Nvector and Vector3d.

New infix and in-place @-operator methods __imatmul__, __matmul__ and __rmatmul__ to convert or transform Cartesian and LatLon instances to a different Datum, Helmert-Transform or RefFrame.

New keyword argument earth to override the datum in Cartesian and LatLon method hartzell.

Documentation updates and several, new cartesian and ellipsoidal tests.