Enhanced Fsum, rev 2.
Fixed imprecise comparison with scalar operations in class Fsum.

Fixed the inconsistent result of method fdiv in class Fsum.

Modified Fsum __ceil__, __floor__, and __divmod__ results to be consistent with Python and return an int quotient in Python 3 but a float in Python 2.

New read-only properties ceil and floor in class Fsum.

Any value to be accumulated in class Fsum methods may be a scalar type or an other Fsum instance.

More Fsum test cases.  Documentation and copyright updates.