CIrcle line intersection
New function intersecant2 in modules sphericalNvector and -Trigonometry and new method intersecant2 for the spherical LatLon classes to compute the intersection of a circle and line.

The circle is specified by its center and either the radius or a point on the circle.  The line is defined as a point inside (or on) the circle and either a bearing or a second point.