Line intersection, wrapped geodesic
New function intersection2 to compute the intersection point of 2 lines, each defined by a lat-, a longitude and a bearing.

New module geodesicw split off from module karney and providing the wrapped Geodesic and GeodesicLine class with leaner error handling.

A few new methods in class GeodesicLineExact and in wrapped classes Geodesic and GeodesicLine plus a fix in the latter.

Limit testresults file names to 200 characters.

Several documentation and test updates.

Update to Python 3.11.3 for building and testing.