Geoids, heights, scypi, triaxials
Updated the geoid and height interpolators to use functions scipy.interpolate.bisplrep and -.bisplev for scipy version 1.14.1 and newer.  For backward compatibility, function scipy.interpolate.interp2d is still used with scipy versions before 1.14.1.

New function triaxum5 to best-fit a triaxial ellipsoid or biaxial ellipse to a set of points.  New function circum4 to best-fit a sphere or circle to a set of points.

Removed localizing globals from function trilaterate3d2 and several other.

Updated to python 3.12.5 and 3.13rc1, to numpy 2.1.0 and scypi 1.14.1 and to macOS 14.6.1 for testing and building.

Documentation and test updates.